+++ test/regex.jl
@@ -32,11 +32,11 @@ show(buf, r"")
 @test read(buf, String) == "r\"\""
 
 # see #10994, #11447: PCRE2 allows NUL chars in the pattern
-@test ismatch(Regex("^a\0b\$"), "a\0b")
+@test contains("a\0b", Regex("^a\0b\$"))
 
 # regex match / search string must be a String
 @test_throws ArgumentError match(r"test", GenericString("this is a test"))
-@test_throws ArgumentError search(GenericString("this is a test"), r"test")
+@test_throws ArgumentError findfirst(r"test", GenericString("this is a test"))
 
 # Named subpatterns
 let m = match(r"(?<a>.)(.)(?<b>.)", "xyz")